pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "agentbench"
version = "0.1.0"
description = "AI Agent evaluation platform for super-individuals"
requires-python = ">=3.10"
dependencies = [
    "openai>=1.0",
    "anthropic>=0.20",
    "python-dotenv>=1.0",
]

[project.scripts]
agentbench = "run:main"
